move test from download manager tests to zip file tests
